requirements.txt
connexion~=2.7.0
flask~=1.1.2
flask-sqlalchemy
jsonschema~=3.2.0
sqlalchemy
swagger-ui-bundle
PyJWT~=1.7.1
markupsafe==2.0.1
flask_cors~= 3.0.10